+++ src/main/java/insane96mcp/progressivebosses/events/EntityRegisterAttributes.java
@@ -1,8 +1,8 @@
 package insane96mcp.progressivebosses.events;
 
 import insane96mcp.progressivebosses.ProgressiveBosses;
-import insane96mcp.progressivebosses.module.dragon.entity.LarvaEntity;
-import insane96mcp.progressivebosses.module.wither.entity.WitherMinionEntity;
+import insane96mcp.progressivebosses.module.dragon.entity.Larva;
+import insane96mcp.progressivebosses.module.wither.entity.WitherMinion;
 import insane96mcp.progressivebosses.setup.PBEntities;
 import net.minecraftforge.event.entity.EntityAttributeCreationEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -12,7 +12,7 @@ import net.minecraftforge.fml.common.Mod;
 public class EntityRegisterAttributes {
 	@SubscribeEvent
 	public static void registerAttributes(EntityAttributeCreationEvent event) {
-		event.put(PBEntities.WITHER_MINION.get(), WitherMinionEntity.prepareAttributes().build());
-		event.put(PBEntities.LARVA.get(), LarvaEntity.prepareAttributes().build());
+		event.put(PBEntities.WITHER_MINION.get(), WitherMinion.prepareAttributes().build());
+		event.put(PBEntities.LARVA.get(), Larva.prepareAttributes().build());
 	}
 }